fix #46
Implement proper handling of quiet option and json output. With this,
some configuration options of the text output like the atom separator
and other small details will be ignored because the default model
printer is no longer used. But in my opinion, the important thing is
that the json output contains all necessary information. All the rest is
cosmetics.